hair_care_project-main/requirements.txt
streamlit
pandas
numpy
scikit-learn
opencv-python-headless
Pillow